Fix bug in exec_while
When exec_while encountered a false in the boolean stack, it always
triggered an exception in interpret-one-step. The problem seemed to be
that leaving the 'else' branch empty in the if function in exec_while
caused a 'nil' to be put on the exec stack. Adding an else branch that
returns the next-next item on the exec stack seemed to do the trick. I
also wrapped the return statement in the 'then' branch in a (seq ...) to
ensure that it is a sequence, which interpret-one-step requires.

I am not convinced exec_while is working correctly, as I'm not sure how
the open/close parentheses work. In the 'opens' def, I attribute 1 open
to exec_while, but it takes two args from the exec stack. I'm not sure
how it handles that.